requirements.txt
python-dotenv==1.0.0
selenium==4.11.2
Pillow==10.0.0
webdriver-manager==4.0.0
